refactor(app): extract terminal style serialization
